adjusted style refactoring
